Fix TicketClassesController#update
Summary:
While adding ticket class attribute changes to logging I noticed that the #update
path is broken. The request test is flawed which exposed the fact that ID is
never passed to the Ticket service so it was only creating.

TicketClassesController now passes params[:id] to the TicketService if it's set.
Ticket Service passes the id value (or nil) to the TicketsApi::Commands
Request tests mock both an ticket class retrieval AND an update.

Test Plan: be rspec

Reviewers: #queue, emilio

Reviewed By: #queue, emilio

Subscribers: emilio

Differential Revision: https://phabricator.evbhome.com/D24813