added sales channels to event extra tickets
Summary:
sales_channels has been added to the EventExtra tickets. This is to support
the marquee being able to show if tickets are "at the door only" or not.

Some tests were updated to actually test the ticket serialization as well as
the EventExtraUpdateJob activity.

Test Plan: be rspec

Reviewers: #queue, vinnie

Reviewed By: #queue, vinnie

Subscribers: vinnie

JIRA Issues: EB-49613

Differential Revision: https://phabricator.evbhome.com/D28890